requirements.txt
aniso8601==9.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
flask-restx==0.5.1
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
